ROX-16734: Refactor Notifiers to pkg
