+++ central/notifier/processor/notifier_set_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stackrox/rox/central/notifiers"
 	"github.com/stackrox/rox/central/notifiers/mocks"
 	"github.com/stackrox/rox/generated/storage"
+	pkgNotifiers "github.com/stackrox/rox/pkg/notifiers"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -87,8 +88,8 @@ func (s *notifierSetTestSuite) TestCoorelatedPoliciesAndNotifiers() {
 	s.mockAlertN.EXPECT().AlertNotify(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 	s.mockResolvableAlertN.EXPECT().AlertNotify(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 
-	s.ns.ForEach(ctx, func(ctx context.Context, n notifiers.Notifier, failures AlertSet) {
-		an, ok := n.(notifiers.AlertNotifier)
+	s.ns.ForEach(ctx, func(ctx context.Context, n pkgNotifiers.Notifier, failures AlertSet) {
+		an, ok := n.(pkgNotifiers.AlertNotifier)
 		if !ok {
 			return
 		}
@@ -98,7 +99,7 @@ func (s *notifierSetTestSuite) TestCoorelatedPoliciesAndNotifiers() {
 	// Check that the resolvable alert notifiers are activated.
 	s.mockResolvableAlertN.EXPECT().AlertNotify(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 
-	s.ns.ForEach(ctx, func(ctx context.Context, n notifiers.Notifier, failures AlertSet) {
+	s.ns.ForEach(ctx, func(ctx context.Context, n pkgNotifiers.Notifier, failures AlertSet) {
 		an, ok := n.(notifiers.ResolvableAlertNotifier)
 		if !ok {
 			return
@@ -109,8 +110,8 @@ func (s *notifierSetTestSuite) TestCoorelatedPoliciesAndNotifiers() {
 	// Check that the audit notifiers are activated.
 	s.mockAuditN.EXPECT().SendAuditMessage(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 
-	s.ns.ForEach(ctx, func(ctx context.Context, n notifiers.Notifier, failures AlertSet) {
-		an, ok := n.(notifiers.AuditNotifier)
+	s.ns.ForEach(ctx, func(ctx context.Context, n pkgNotifiers.Notifier, failures AlertSet) {
+		an, ok := n.(pkgNotifiers.AuditNotifier)
 		if !ok {
 			return
 		}
