+++ central/notifiers/metadatagetter/singleton.go
@@ -1,7 +1,7 @@
 package metadatagetter
 
 import (
-	"github.com/stackrox/rox/central/notifiers"
+	"github.com/stackrox/rox/pkg/notifiers"
 	"github.com/stackrox/rox/pkg/sync"
 )
