+++ pkg/notifiers/syslog/mocks/syslog.go
