+++ pkg/notifiers/syslog/tcp_sender.go
